Filter sequences by full name/description instead of id
Switches from faSomeRecords to seqkit grep in order to do so.

faSomeRecords only matches on the FASTA id¹, so filtering for

    hCov-19/South Africa/…

will also include sequences like

    hCov-19/South Korea/…

This repo's sequence selection list (data/strains_africa.txt) contains
the full FASTA name/description for each sequence (as id isn't unique in
this data), so using faSomeRecords is not appropriate.

seqkit grep also defaults to matching on the id just like faSomeRecords,
but the --by-name option changes it to match on the full
name/description.  That's what we need.

Avoiding faSomeRecords is also useful because, unlike seqkit, it's not
in our Nextstrain runtimes.

¹ The id is the part between the initial ">" and the first space, e.g.
  in ">x y z" the id is "x" and the full name/description is "x y z".

Related-to: <https://github.com/nextstrain/ncov-africa-cdc/pull/7>

Co-authored-by: bioaddict <rafgangbadja@gmail.com>
Co-authored-by: Thomas Sibley <tsibley@fredhutch.org>